Rework the card styling for a neater look
Modified the color scheme of cards to match the app's background schema
as the original yellow cards did not match well with the overall new
dark look of the app.

Added hover styles to cards in order to create a intuitive user
experience.

Adjusted the image link added to the card prop in order to effect
changes in the app's UI as well as to follow standard link convention
on the Unsplash API.