Improve accessibility of the web app
Included a main landmark for a better page structure.

Added empty alt attributes to images to prevent confusing screen-readers
of the purpose of the rendered images. Alt text was not needed since the
images already have captions attached to them.

Modify cards to use buttons instead of divs to indicate the intention
for interaction to screen readers.